ref: Move bundle into apm package
